Refactoring Round 1 (#228)
## Changes:

- add pre-commit with steps for ruff, trailing whitespaces, end of files
and double quotes
- add "checker" pipeline to automatically run pre-commit (with ruff),
mypy and pylint on the entire code base
- add `requirement-dev.txt` to manage development requirements
- add pyproject.toml to save configurations for mypy, pylint, pytest,
ruff and the project definition
- correct usage of cache with custom user in the Dockerfile
- only import classes and functions to `__init__.py`s that are part of
the public API of that module
- remove `typing.Optional` and `typing.Union` in favor of `|`
- correct imports:
- `from system import System` -> `from ..system import System` (works,
but IDE is confused)
- `from __future__ import annotations` -> to properly use `if
TYPE_CHECKING` imports (`System` instead of `'System'`)
  - remove unused imports, reorder imports according to ruff
- import from neighboring files (`from .field import Field, Row` ->
directly from file instead of `from . import Field, Row` -> relying on
`__init__.py`)
- typing:
- making sure optional variables are really set and/or properly treat if
they are not
- when a variable can be different types, make sure that the called
function exists
  - make arguments keyword arguments if 5 or more
- refactor `header.py` and other ui elements containing header, drawer
and footer
- make functions private that are only used internally in a class
- rename classes to use PascalCase, functions that are creating ui and
are exported to `create_xyz_ui`
- refactor ui components using `ui.markdown` -> `ui.label`
- use `PackagePathFilter` of RoSys to log paths
- remove `coverage_planner.py`, `chop_and_screw.py`, `path_recorder.py`,
`sequence.py`, `path_planner.py`, `components/test.py`, `pages` folder,
`point_transformation.py`

## Open Points

- [ ] refactor `localization` module and find different solution to
`localization.reference` import (cyclic import)
- [ ] refactor flashlight classes into one or two at maximum
- [ ] refactor sections with code duplication (`camera_card.py`,
`status_dev`,
- [ ] new Exceptions (`NotReferencedException`, `MoveImplementException`
(out of work area, could not move, ...))
- [ ] refactor GeoDataImporter
- [ ] refactor ui components with long update function (use bindings or
`module_ui` function)
- [ ] check tests with mypy and pylint

---------

+++ field_friend/vision/calibration.py
@@ -1,6 +1,8 @@
-import logging
+# pylint: disable=no-member
+# TODO: how to get better type checking for cv2?
+from collections.abc import Sequence
 from dataclasses import dataclass
-from typing import Self, Sequence
+from typing import Self
 
 import cv2
 import numpy as np
@@ -31,7 +33,8 @@ class Contour:
         shape_similarity = cv2.matchShapes(self.points, ellipse_contour, cv2.CONTOURS_MATCH_I3, 0.0)
         return shape_similarity < shape_threshold
 
-    def fit_ellipse(self) -> tuple[Sequence[float], Sequence[int], float]:
+    # TODO: return type was tuple[Sequence[float], Sequence[int], float] <- do we now break things with the new one
+    def fit_ellipse(self) -> tuple[Sequence[float], Sequence[float], float]:
         return cv2.fitEllipse(self.points)
 
     def contains(self, x: float, y: float) -> bool:
@@ -91,7 +94,8 @@ class Network:
         return network
 
     def try_refine(self, dot: Dot, *, tolerance: float = 0) -> None:
-        for dx in np.arange(0, tolerance + 1, 10):
+        # TODO: can we get rid of the pylint disable? -> no, we live with it, but maybe document it?
+        for dx in np.arange(0, tolerance + 1, 10):  # pylint: disable=too-many-nested-blocks
             for dy in np.arange(0, tolerance + 1, 10):
                 for sign_x in [-1, 1] if dx else [0]:
                     for sign_y in [-1, 1] if dy else [0]:
